Extract the Secrets module (secrets at rest) into Api/Secrets
Second module of the #777 folder migration (option B — folders with matching
namespaces). Moves the at-rest secret handling into
Jellyfin.Plugin.SSO_Auth.Api.Secrets:

- SecretStore, SecretEnvelope, ConfigSecretProtection -> Api/Secrets/
  (namespace Api.Secrets), import added at each call site.

ConfigSecretProtection keeps its dependency on the Config persistence model
(SamlConfig signing-key fields) — a non-Api namespace, which the leaf rule
allows; the two crypto primitives depend on nothing but the BCL.

Generalises the module-boundary fitness function into a parameterised theory
ApiModule_IsALeaf_ImportsNoOtherApiModule (Net, Secrets) instead of duplicating
the per-module body, so each further module adds one InlineData case. The rule
now also ignores a module's self-import defensively.

Pure mechanical move: no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1573 both TFMs), every
existing conformance rule (incl. the Secret<T>-redaction / secrets structural
rules), and the ABI-floor build.

Refs #777
