+++ SSO-Auth.Tests/SecretEnvelopeTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Security.Cryptography;
 using Jellyfin.Plugin.SSO_Auth.Api;
+using Jellyfin.Plugin.SSO_Auth.Api.Secrets;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
